Removed calculations from ConstrictorAttribute
