+++ Constrictor/Constrictor/Extensions/Constrictable+ConstrictorCore.swift
@@ -31,8 +31,8 @@ extension Constrictable {
                    to item: Constrictable, attribute: ConstrictorAttribute, constant: Constant,
                    multiplier: CGFloat = 1.0, priority: UILayoutPriority = .required) {
         
-        let firstItemLayoutAttribute = selfAttribute.itemLayoutAttribute(for: self, with: constant)
-        let secondItemLayoutAttribute = attribute.itemLayoutAttribute(for: item, with: constant)
+        let firstLayoutAttributes = ItemLayoutAttributesDecoder.itemLayoutAttribute(for: self, with: selfAttribute, and: constant)
+        let secondLayoutAttributes = ItemLayoutAttributesDecoder.itemLayoutAttribute(for: item, with: attribute, and: constant)
 
         if let constrictableAsView = self as? UIView {
 
@@ -41,12 +41,12 @@ extension Constrictable {
 
         
         NSLayoutConstraint(item: self,
-                           attribute: firstItemLayoutAttribute.layoutAttribute,
+                           attribute: firstLayoutAttributes.layoutAttribute,
                            relatedBy: relation,
-                           toItem: secondItemLayoutAttribute.item,
-                           attribute: secondItemLayoutAttribute.layoutAttribute,
+                           toItem: secondLayoutAttributes.item,
+                           attribute: secondLayoutAttributes.layoutAttribute,
                            multiplier: multiplier,
-                           constant: firstItemLayoutAttribute.constant).isActive = true
+                           constant: firstLayoutAttributes.constant).isActive = true
     }
     
     /**
@@ -67,7 +67,7 @@ extension Constrictable {
     func constrict(_ selfAttribute: ConstrictorAttribute, relation: NSLayoutRelation = .equal,
                    constant: Constant, multiplier: CGFloat = 1.0, priority: UILayoutPriority = .required) {
         
-        let layoutAttributes = selfAttribute.itemLayoutAttribute(for: self, with: constant)
+        let layoutAttributes = ItemLayoutAttributesDecoder.itemLayoutAttribute(for: self, with: selfAttribute, and: constant)
 
         if let constrictableAsView = self as? UIView {
